requirements.txt
Django==4.2
djangorestframework==3.14.0
mysqlclient==2.2.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
pyjwt==2.8.0
python-dotenv==1.0.0
django-cors-headers==4.3.0
pandas==2.1.1
openpyxl==3.1.5
